% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biomart_hsapiens}
\alias{biomart_hsapiens}
\title{Default mapping for entrezgene IDs and HGNC gene symbols}
\format{
A data.frame containing a mapping between entrezgene IDs and
HGNC gene symbols.
}
\usage{
biomart_hsapiens
}
\description{
This dataset is used by default if the connection to biomaRt fails. It
is highly recommended to retry the function call that attempted to connect to
biomaRt. Using this dataset in general may not produce the correct
results.
}
\keyword{datasets}
