% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{c.dnapath}
\alias{c.dnapath}
\title{Combine two 'dnapath' objects.}
\usage{
\method{c}{dnapath}(...)
}
\arguments{
\item{...}{'dnapath' objects to be concatenated.}
}
\value{
Concatenation is not defined; an error is generated.
}
\description{
This functionality is not implemented and will return an error.
}
