% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{c.dnapath_list}
\alias{c.dnapath_list}
\title{Combine two 'dnapath_list' objects.}
\usage{
\method{c}{dnapath_list}(...)
}
\arguments{
\item{...}{'dnapath_list' objects to be concatenated.}
}
\value{
Concatenation is not defined; an error is generated.
}
\description{
This functionality is not implemented and will return an error.
}
