% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_edgesC}
\alias{d_edgesC}
\title{C++ implementation of d_edges}
\usage{
d_edgesC(nw1, nw2, lp)
}
\arguments{
\item{nw1}{The association scores for network 1}

\item{nw2}{The association scores for network 2}

\item{lp}{The lp value to use.}
}
\value{
A matrix of differential network scores for the edges.
}
\description{
Calculates differential network score for each edge in a network
}
