% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_genesC}
\alias{d_genesC}
\title{C++ implementation of d_genes}
\usage{
d_genesC(nw1, nw2, lp)
}
\arguments{
\item{nw1}{The association scores for network 1}

\item{nw2}{The association scores for network 2}

\item{lp}{The lp value to use.}
}
\value{
A vector of differential network scores for the genes.
}
\description{
Calculates differential network score for a set of genes
}
