% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_pathwayC}
\alias{d_pathwayC}
\title{C++ implementation of d_pathway}
\usage{
d_pathwayC(nw1, nw2, lp)
}
\arguments{
\item{nw1}{The association scores for network 1}

\item{nw2}{The association scores for network 2}

\item{lp}{The lp value to use.}
}
\value{
The differential network score for the pathway.
}
\description{
Calculates differential network score for an entire pathway.
}
