% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome.R
\name{format_species_name}
\alias{format_species_name}
\title{Format sepcies name input.}
\usage{
format_species_name(species)
}
\arguments{
\item{species}{The species to obtain a biomart dataset for. For example:
"Homo sapiens", "m musculus", "C. elegans", or "S cerevisiae", and
"Human" and "mouse" will be converted to the format required in biomart.}
}
\value{
A string containing the formatted species name.
}
\description{
Internal function used to format species names to be used with biomart.
}
\keyword{internal}
