% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{get_mean_expr_mat}
\alias{get_mean_expr_mat}
\title{Get mean expression of pathway genes in both groups}
\usage{
get_mean_expr_mat(x)
}
\arguments{
\item{x}{A 'dnapath' or 'dnapath_list' object.}
}
\value{
A matrix of two rows whose columns contain the mean expression of
each gene in the pathway.
}
\description{
Get mean expression of pathway genes in both groups
}
\keyword{internal}
