% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{head.dnapath_list}
\alias{head.dnapath_list}
\title{Return the first part of the dnapath results.}
\usage{
\method{head}{dnapath_list}(x, ...)
}
\arguments{
\item{x}{A 'dnapath_list' object.}

\item{...}{Additional paramters are passed into
\code{\link{summary.dnapath_list}}.}
}
\value{
Returns the first five rows of the summary table of the
'dnapath_list' object.
}
\description{
Return the first part of the dnapath results.
}
\examples{
data(meso)
data(p53_pathways)
set.seed(0)
results <- dnapath(x = meso$gene_expression, pathway_list = p53_pathways,
                   group_labels = meso$groups, n_perm = 10)
head(results)
}
