% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome.R
\name{init_mart}
\alias{init_mart}
\title{Initialize biomaRt for mmusculus}
\usage{
init_mart(species)
}
\value{
None. Creates a global variable called "mart" used by the
\code{\link{symbol_to_entrez}} and \code{\link{entrez_to_symbol}} methods.
}
\description{
Initialize biomaRt for mmusculus
}
\keyword{internal}
