% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{length.dnapath_list}
\alias{length.dnapath_list}
\title{The number of pathways in a 'dnapath_list' object.}
\usage{
\method{length}{dnapath_list}(x)
}
\arguments{
\item{x}{A 'dnapath_list' object from \code{\link{dnapath}}.}
}
\value{
The number of pathways.
}
\description{
The number of pathways in a 'dnapath_list' object.
}
\examples{
data(meso)
data(p53_pathways)
set.seed(0)
results <- dnapath(x = meso$gene_expression, pathway_list = p53_pathways,
                   group_labels = meso$groups, n_perm = 10)
length(results)
}
