% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mart}
\alias{mart}
\title{Default mart obtained from biomaRt for H. sapiens}
\format{
A 'Mart' object from biomaRt.
}
\usage{
mart
}
\description{
This dataset is used by default if the connection to biomaRt fails. It
is highly recommended to retry the function call that attempted to connect to
biomaRt. Using this dataset in general will not produce the correct
results.
}
\keyword{datasets}
