% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meso}
\alias{meso}
\title{Gene expression dataset for two groups}
\format{
A list containing two items:
\describe{
\item{$gene_expression}{A 32 by 150 matrix of gene expression values}
\item{$groups}{A vector of length 32 indicating which group (stageii or
stageiv) each of the rows in the gene expression data belong to.}
}
}
\source{
\url{http://www.linkedomics.org/data_download/TCGA-GBMLGG/}
}
\usage{
meso
}
\description{
\code{meso} is a list containing gene expression data from Mesothelioma tumors
generated by The Cancer Genome Atlas (TCGA) and obtained using the
LinkedOmics portal. The first element in the list, named "gene_expression",
contains 32 samples (rows) with 150 genes (columns). The second element,
named "groups", is a vector of length 32 indicating which group (stage ii
or stage iv) each gene expression sample belongs to. See the "Package data"
vignette for details.
}
\keyword{datasets}
