% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{names.dnapath}
\alias{names.dnapath}
\title{The pathway names in a 'dnapath' object.}
\usage{
\method{names}{dnapath}(x)
}
\arguments{
\item{x}{A 'dnapath' object from \code{\link{dnapath}} or from subsetting
a 'dnapath_list'.}
}
\value{
The pathway's name.
}
\description{
The pathway names in a 'dnapath' object.
}
\examples{
data(meso)
data(p53_pathways)
set.seed(0)
results <- dnapath(x = meso$gene_expression, pathway_list = p53_pathways,
                   group_labels = meso$groups, n_perm = 10)
names(results[[1]])
}
