% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p53_pathways}
\alias{p53_pathways}
\title{Reactome pathway list for Homo sapiens}
\format{
A list of 13 vectors each containing a set of entregene IDs.
}
\usage{
p53_pathways
}
\description{
This is a pathway list obtained from \code{\link{get_reactome_pathways}}
with \code{species = "human"} (used \code{reactome.db} version 1.68.0).
Only pathways with "p53" in their name are retained (to subset on some
cancer-related pathways). The list contains 13 total pathways. See the
"Package data" vignette for details.
}
\keyword{datasets}
