% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{print.dnapath}
\alias{print.dnapath}
\title{Print function for 'dnapath' object.}
\usage{
\method{print}{dnapath}(x, ...)
}
\arguments{
\item{x}{A 'dnapath' object from \code{\link{dnapath}}.}

\item{...}{Additional arguments are ignored.}
}
\value{
Prints a summary of the module.
}
\description{
Print function for 'dnapath' object.
}
