% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{[.dnapath_list}
\alias{[.dnapath_list}
\title{Extract parts of a 'dnapath_list' object.}
\usage{
\method{[}{dnapath_list}(x, i, ...)
}
\arguments{
\item{x}{A 'dnapath_list' object from \code{\link{dnapath}}.}

\item{i}{The indices of pathways to extract.}

\item{...}{Additional arguments are ignored.}
}
\value{
A 'dnapath_list' object containing pathways indexed by 'i'.
}
\description{
Extract parts of a 'dnapath_list' object.
}
