% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{[<-.dnapath}
\alias{[<-.dnapath}
\title{Replace parts of a 'dnapath' object.}
\usage{
\method{[}{dnapath}(x, ...) <- value
}
\arguments{
\item{x}{A 'dnapath' object from \code{\link{dnapath}}.}

\item{...}{Additional arguments are ignored.}

\item{value}{A 'dnapath' object.}
}
\value{
Replacement is not defined; an error is generated.
}
\description{
This functionality is not implemented and will return an error.
}
