% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna.R
\name{[<-.dnapath_list}
\alias{[<-.dnapath_list}
\title{Replace parts of a 'dnapath_list' object.}
\usage{
\method{[}{dnapath_list}(x, ...) <- value
}
\arguments{
\item{x}{A 'dnapath_list' object from \code{\link{dnapath}}.}

\item{...}{Additional arguments are ignored.}

\item{value}{A 'dnapath_list' object.}
}
\value{
Replacement is not defined; an error is generated.
}
\description{
This functionality is not implemented and will return an error.
}
