% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_edges}
\alias{summarize_edges}
\title{Summarize differential connections for a pathway}
\usage{
summarize_edges(x, alpha_edge = NULL, monotonized = FALSE)
}
\arguments{
\item{x}{A 'dnapath' object from \code{\link{dnapath}}.}

\item{alpha_edge}{Threshold for p-values of edge DC scores. If NULL,
defaults to 0.05 or the minimum possible threshold (based on the
number of permutatiosn that were run).}

\item{monotonized}{If TRUE, monotonized p-values are used.}
}
\value{
A tibble summarizing the differential connections in
the pathway.
}
\description{
Summarize differential connections for a pathway
}
\examples{
data(meso)
data(p53_pathways)
set.seed(0)
results <- dnapath(x = meso$gene_expression, pathway_list = p53_pathways,
                   groups = meso$groups, n_perm = 10)
summarize_edges(results[[1]])
}
\seealso{
\code{\link{summarize_pathways}}, \code{\link{summarize_genes}}
}
