% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize_genes_for_pathway}
\alias{summarize_genes_for_pathway}
\title{Summarize differential connectivity of genes in a pathway}
\usage{
summarize_genes_for_pathway(x, alpha = NULL, monotonized = FALSE)
}
\arguments{
\item{x}{A 'dnapath' object of a single pathway.}

\item{alpha}{Threshold for p-values of gene DC scores, used to
subset the x. If NULL (or 1), x for all genes are shown.}

\item{monotonized}{If TRUE, monotonized p-values are used.}
}
\value{
A tibble summarizing the differential connectivity of genes in
the pathway.
}
\description{
Summarize differential connectivity of genes in a pathway
}
\keyword{internal}
