% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactome.R
\name{init_mart}
\alias{init_mart}
\title{Initialize biomaRt for a given species}
\usage{
init_mart(species)
}
\arguments{
\item{species}{The species to obtain a biomart dataset for. This should be
an output from \code{\link{format_species_name}}.}
}
\value{
A biomaRt object.
}
\description{
Initialize biomaRt for a given species
}
\note{
Requires internet connection. This function is called by
\code{\link{get_biomart_mapping}}, which checks for internet connection
prior to calling this function.
}
\keyword{internal}
